Fixed a bug in the sampling behavior
The sampling procedure previously enforced periodic BC's on the input
coordinates. This turned out to be problematic when the sample was very
near the domain boundary, causing the result of indexatposition to be
Nx_global + 2*Ng - 1 instead of Ng, which is way out of bounds.

This commit changes the logic of the setsamplecoords function to reject
a set of samples containing an off-bounds point, having a null effect on
the data structure and returning an error code other than zero.