Fixes issue in fromfile withguard zones
When the fromfile function is called with guard zones, the h5py reads
need to go to the interior of the array only. Lines in the test script
demonstrate that this works. When h5py is used to read the file, guard
zones must be synced, so this commit also adds that method to the
DataField class.