MAJOR FIX: Pair sampling mystery solved
Bins at large spatial separation are left severely under populated when
the output list of coordinates is sampled sequentially. This is due to
the fact that the point sampling routine returns the input coordinates
organized by subgrid, and thus adjacent points in the list are at close
proximity. Any correct usage of the sampling function must randomize its
selection of points from the output list.