Implements the sampling functions in Cython
This was slightly more involved. Not complicated, just a few extra
lines. There's no special Cython syntax equivalent to numpy.i's
ARGOUTVIEW_ARRAY. To get a copy of the data represented in a C data
structure, you have to get the pointer to the data, then copy it
manually into a Numpy array.