Introduces higher level structure function calculations
There is a new function in the srhdpack.c file called
srhdpack_collectpairs which returns sample values to later be added into
histograms. This method allows any number of processing modes to be
passed in via an array of struct srhdpack_samplemode, each of which
applies a different post-processing to every pair taken at the C
level. The processed pairs are returned as 2d arrays with the mode on
the second axis.

Python code passes sanity checks on running the low-level C
function. The older version of the function has not been removed from
the code yet, give it some time.